replace autoload in a post spl universe
The old autoload was a system that would emulate various parts of TAGGED's old codebase in a pre-PHP 5.1 universe. All of that stuff is outdated post spl_autoload_register.

1. classmap tables for backward compatibility can be handled easily in composer
2. applications based on a framework register their own autoload handler
3. PSR-0 or PSR-4 compliance means we can register our own autoloader

Note that (3) is not complete yet because a lot needs to be migrated for PSR-4 (namespaces rewritten), as well as class names were strtolowered non-case sensitive class naming conventions/shortcut (class names were not case sensitive) which breaks PSR-1. So there is still an autoload function being registered (for now).

Signed-off-by: terry chay <tychay@php.net>